[🚀 사이클2 - 미션 (예약 변경/취소와 에러 처리)] 루드비코 미션 제출합니다. (#482)
* refactor: 도메인 객체의 equals and hashCode 재정의를 제거

현재 도메인 객체는 id를 포함하는 엔티티로 사용되고 있다. 모든 필드가
같을 때 동등한 객체로 판단하는 것은 value object에서 필요한 조건이다.

따라서 엔티티가 value object처럼 사용되는 위험을 제거한다.

이에 따라 테스트에서 isEqualTo가 아닌, usingRecursiveComparsion 등으로
내부 필드값으로 비교하도록 한다.

* docs: 잘못된 요청 거부 요구사항 추가

* feat: 이미 지난 시점에 대한 예약 요청 검증 추가

* feat: 비어 있는 이름에 대한 예약 요청 검증 추가

* refactor: 과거 시점 예약에 대한 예외 메시지의 날짜, 시간 출력 포맷 변경

기존에 사용하던 ISO_DATE는 날짜만 출력되는 문제가 있다. ISO_DATE_TIME은
시간에 초단위 정보 등 불필요한 정보가 추가 노출된다.

대신 커스텀 패턴을 적용하여 필요한 정보만 읽기 쉽게 출력되도록 한다.

* feat: 예약이 존재하는 예약 시간에 대한 삭제 요청 검증 추가

* refactor: 날짜, 시간, 테마에 대한 중복된 예약 요청의 응답 변경

기존 500 응답에서 409로 변경한다.

* refactor: 잘못된 날짜 형식에 대한 예약 요청의 응답을 변경

* refactor: 지난 시점에 대한 예약 요청의 응답을 변경

* refactor: 비어 있는 이름에 대한 예약 요청 응답 변경

* refactor: 예외 핸들링 테스트 책임을 컨트롤러에서 예외 핸들러로 이동

* refactor: 컨트롤러 테스트를 E2E 테스트로 변경

* refactor: 과거 시점에 예약에 대한 예외가 다음 행동을 유도하도록 변경

* refactor: 예약에 대한 RestAssured 테스트를 컨트롤러 E2E 테스트로 이동

* refactor: 예약 시간에 대한 RestAssured 테스트를 컨트롤러 E2E 테스트로 이동

* refactor: 테마에 대한 RestAssured 테스트를 컨트롤러 E2E 테스트로 이동

* feat: 사용자별 예약 목록 조회 데이터 접근 로직 구현

* feat: 사용자별 예약 목록 조회 비즈니스 로직 구현

* feat: 사용자별 예약 목록 조회 API 구현

* feat: 사용자의 본인 예약 취소 데이터 접근 로직 구현

* feat: 사용자의 본인 예약 취소 비즈니스 로직 구현

* feat: 사용자의 본인 예약 취소 API 구현

* feat: 사용자의 본인 예약 날짜, 시간 변경 데이터 접근 로직 구현

* feat: 사용자의 본인 예약 날짜, 시간 변경 비즈니스 로직 구현

* feat: 사용자의 본인 예약 날짜, 시간 변경 API 구현

* feat: 본인 예약 조회/변경/취소 기능을 web UI에 반영

* refactor: java 코드에서의 sql 키워드를 대문자로 변경

* refactor: 파라미터가 3개 이상인 쿼리문 실행에 NamedParameterJdbcTemplate 사용

* refactor: 예외 계층 구조화 및 ErrorResponse 도입

* refactor: 예외 핸들러를 전역과 비즈니스 범위로 분리

* feat: 커스텀 예외 세분화 및 예외 응답 메시지 친절하게 만들기

에러 메시지에 대해 다음 원칙을 적용한다.
- 현재 상태와 원인, 해결 방법을 알려줘야 한다.
- 예시: 저장 공간이 부족하여 파일을 업로드하지 못했습니다.
저장 공간을 확보한 후 다시 시도해 주세요.

이때 각 상태와 원인을 구분하기 위해 커스텀 예외를 세분화한다. 예를 들어
기존에는 예약 취소 실패에 대한 예외가 하나였지만, 이를 존재하지 않는
예약, 과거 예약 등으로 세분화한다.

* refactor: 뷰 템플릿에서 API가 내려주는 구조화된 에러 응답 정보를 사용

* refactor: 코드 스타일 정리 및 unused import 제거

* refactor: 전체 시간 목록 캐싱 및 날짜 변경 시 예약된 시간만 요청하도록 개선

전체 시간 슬롯은 항상 동일한 정보를 응답하기 때문에 js 수준에서 이를
캐싱하도록 한다.

* build: add spring-boot-starter-validation dependency

* refactor: 사용자 이름으로 예약 조회 API가 검증 실패 시 400 예외를 던지도록 수정

* fix: 예약 조회 시 사용자 이름 누락에 대한 방어 로직 추가

* refactor: 요청 필수값 검증 및 위반 시 400 응답하도록 변경

* refactor: 미사용 엔드포인트 제거

* refactor: 리포지토리 메서드명을 가독성 좋게 변경

- 메서드의 행위를 구체적으로 명시한다(찾는다 -> 예약을 찾는다)
- 조건문이 많을 때 메서드명에 나열하지 않는다(byTimeIdAnd... -> with)

* refactor: 서비스 메서드명을 비지니스 케이스에 가깝게 변경

* test: 누락된 ExceptionHandler 테스트 추가

* refactor: 업데이트 하려는 값이 기존과 같을 때 409 대신 204를 응답

* refactor: 비지니스 상황의 Not Found를 404 대신 422로 응답

* docs: 기능 명세 및 에러 응답 명세 추가

* refactor: 누락된 DTO 검증 추가

* refactor: 예약이 존재하는 테마 삭제 시 500 대신 422 응답

* feat: 도메인 레벨 검증 로직 추가

* refactor: 서비스 계층에서 도메인 계층 검증 로직을 사용

* docs: 에러 응답 예시 추가

* test: 관리자 예약 조회 및 테마 상세 조회 E2E 테스트 추가

* refactor: 잘못된 타입으로 API 요청에 대한 응답을 500에서 400으로 변경

* refactor: 오타 수정

* refactor: 존재 여부를 묻는 메서드명 오타 수정(exist... -> exists...)

* test: 누락한 E2E 테스트 추가

* refactor: json 역직렬화 오류 시 500 대신 400을 응답

+++ src/main/java/roomescape/service/ThemeService.java
@@ -6,7 +6,9 @@ import org.springframework.transaction.annotation.Transactional;
 import roomescape.domain.Theme;
 import roomescape.dto.ThemeRequestDTO;
 import roomescape.dto.ThemeResponseDTO;
-import roomescape.repository.JdbcThemeRepository;
+import roomescape.exception.ThemeInUseException;
+import roomescape.exception.ThemeNotFoundException;
+import roomescape.repository.ReservationRepository;
 import roomescape.repository.ThemeRepository;
 
 @Service
@@ -14,9 +16,11 @@ import roomescape.repository.ThemeRepository;
 public class ThemeService {
 
     private final ThemeRepository themeRepository;
+    private final ReservationRepository reservationRepository;
 
-    public ThemeService(JdbcThemeRepository themeRepository) {
+    public ThemeService(ThemeRepository themeRepository, ReservationRepository reservationRepository) {
         this.themeRepository = themeRepository;
+        this.reservationRepository = reservationRepository;
     }
 
     public ThemeResponseDTO addTheme(ThemeRequestDTO request) {
@@ -35,7 +39,7 @@ public class ThemeService {
 
     public ThemeResponseDTO findById(Long id) {
         Theme result = themeRepository.findById(id)
-                .orElseThrow();
+                .orElseThrow(() -> new ThemeNotFoundException("ID로 테마 조회 실패: " + id));
         return ThemeResponseDTO.from(result);
     }
 
@@ -47,6 +51,12 @@ public class ThemeService {
     }
 
     public void deleteTheme(Long id) {
+        if (reservationRepository.existsReservationByThemeId(id)) {
+            throw new ThemeInUseException("테마 삭제 실패 (사용 중): " + id);
+        }
+        if (themeRepository.findById(id).isEmpty()) {
+            throw new ThemeNotFoundException("테마 삭제 실패 (존재하지 않음): " + id);
+        }
         themeRepository.delete(id);
     }
 }
